requirements.txt
# Core
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
scikit-surprise==1.1.3
imbalanced-learn==0.11.0

# Deep Learning & NLP
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
datasets==2.15.0
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0.tar.gz

# Vector Search & DB
faiss-cpu==1.7.4  # Com suporte a GPU
chromadb==0.4.22
langchain==0.1.0

# Visualization & UI
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
jupyter==1.0.0
ipython==8.18.1

# Data Access
kaggle==1.6.12
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Development
black==23.11.0
flake8==6.1.0
pytest==7.4.3